Use the user ID as username when it cannot be retrieved
There are cases where the database query will not return a username:
when the uploader name was oversighted (and presumably revdel-d as well?)

In this case, let's bind the user id as the username.

This has the benefit that the files will still be counted,
and correctly mapped to a single user.

The registration data will be off − let's revisit later.

Bug: T264217