Remove category exception from Armenia
Back in e810058, we added category exception for Armenia
to instead target `Images from Wiki Loves Monuments in Armenia & Nagorno-Karabakh`,
as this was the pattern used for WLM 2015 and 2017 as well as WLE 2014.

It turned out this naming convention only held for these three instances,
and we had to add a specialy exception every year since then

As it's now clear what is considered the default pattern,
let's revert this and instead add a handful of exceptions
for these three instances.

Reverts e810058

Also reverts:
- a488a29 Add an exception for Wiki Loves Monuments 2023 in Armenia
- 19b5d50 Add special exception for Wiki Loves Earth 2023 in Armenia
- e965b41 Add an exception for Wiki Loves Monuments 2022 in Armenia. (#39)
- 23a61d2 Add an exception for Wiki Loves Monuments 2021 in Armenia. (#22)
- aebc1f9 Add special exception for WLE 2021 in Armenia
- dfbe5ea Add special exception for WLM 2020 in Armenia
- a8dd9ea Add special exception for WSC 2015 in Armenia
- b13167a Add special exception for WLE 2020 in Armenia
- 9b1d32b Add special exception for WLM 2019 in Armenia

Bug: T364035