Make database file update atomic
There are currently issues with wikiloves,
with the webserver being inaccessible.

At startup time, the web application tries
loading the JSON database in memory.

That JSON database is updated in-place by the update process,
which runs every 15 minutes during a competition.

It is thus possible that the web process tries
reading the file while it’s halfway written,
which would explain some errors in the logs [1].

Here we instead write the database as a separate JSON file,
and copy it in place once it’s done.

[1] Stack trace:
```
Traceback (most recent call last):
  File "app.py", line 50, in <module>
    loadDB()
  File "app.py", line 41, in loadDB
    db = json.load(f)
  File "/usr/lib/python3.9/json/__init__.py", line 293, in load
    return loads(fp.read(),
  File "/usr/lib/python3.9/json/__init__.py", line 346, in loads
    return _default_decoder.decode(s)
  File "/usr/lib/python3.9/json/decoder.py", line 337, in decode
    obj, end = self.raw_decode(s, idx=_w(s, 0).end())
  File "/usr/lib/python3.9/json/decoder.py", line 355, in raw_decode
    raise JSONDecodeError("Expecting value", s, err.value) from None
json.decoder.JSONDecodeError: Expecting value: line 1 column 1 (char 0)
```