/images: lazy load images
Without lazy loading the images, the browser will try to load all of the potential 200 images at once. If there is a lot of images on the page this will trigger Wikimedia's rate limit causing some images to not render.

loading="lazy"  will ensure that images are loaded as the user scrolls rather than all at once.